+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## version 0.0.0a16 29 August 2026
+1. Implement ButtonFrame locally for icon system
+2. tidy main.py for module caller
+3. tidy import statements
+
 ## version 0.0.0a15 28 August 2026
 1. Remove frm_xxx